[SERVICE-460] WIP
