+++ src/client/main.ts
@@ -23,28 +23,7 @@ export * from './contextChannels';
 export * from './context';
 export * from './directory';
 export * from './intents';
-
-export enum OpenError {
-    AppNotFound = 'AppNotFound',
-    ErrorOnLaunch = 'ErrorOnLaunch',
-    AppTimeout = 'AppTimeout',
-    ResolverUnavailable = 'ResolverUnavailable'
-}
-
-export enum ResolveError {
-    NoAppsFound = 'NoAppsFound',
-    ResolverUnavailable = 'ResolverUnavailable',
-    ResolverTimeout = 'ResolverTimeout',
-    InvalidContext = 'InvalidContext'
-}
-
-export class FDC3Error extends Error{
-    public code: string;
-    public constructor(code: string, message: string) {
-        super(message);
-        this.code = code;
-    }
-}
+export * from './errors';
 
 /**
  * Intent descriptor
